+++ apps/backend/src/modules/comments/comment.controller.ts
@@ -1,4 +1,7 @@
+import { type FastifyRequest } from "fastify";
+
 import { APIPath } from "~/libs/enums/enums.js";
+import { checkIdEquality } from "~/libs/hooks/hooks.js";
 import {
 	type APIHandlerOptions,
 	type APIHandlerResponse,
@@ -14,6 +17,7 @@ import {
 	type CommentCreateRequestDto,
 	type CommentGetAllRequestDto,
 	type CommentUpdateRequestDto,
+	type CommentWithRelationsResponseDto,
 } from "./libs/types/types.js";
 import {
 	commentCreateBodyValidationSchema,
@@ -120,6 +124,35 @@ class CommentController extends BaseController {
 				params: commentIdParameterValidationSchema,
 			},
 		});
+
+		this.addRoute({
+			handler: (options) => {
+				return this.delete(
+					options as APIHandlerOptions<{
+						params: Record<"id", number>;
+					}>,
+				);
+			},
+			method: "DELETE",
+			path: CommentsApiPath.$ID,
+			preHandlers: [
+				checkIdEquality<
+					CommentWithRelationsResponseDto,
+					FastifyRequest & {
+						params: Record<"id", number>;
+						user: UserAuthResponseDto;
+					}
+				>({
+					pathInDtoToCompareId: "userId",
+					pathInRequestToCompareId: "user.id",
+					pathInRequestToDtoId: "params.id",
+					service: this.commentService,
+				}),
+			],
+			validation: {
+				params: commentIdParameterValidationSchema,
+			},
+		});
 	}
 
 	/**
@@ -166,6 +199,44 @@ class CommentController extends BaseController {
 				...options.body,
 			}),
 			status: HTTPCode.CREATED,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /comments/{id}:
+	 *    delete:
+	 *      tags:
+	 *        - Activity Comments
+	 *      description: Delete own comment by provided comment id
+	 *      security:
+	 *        - bearerAuth: []
+	 *      parameters:
+	 *        - name: id
+	 *          in: path
+	 *          description: The comment id
+	 *          required: true
+	 *          schema:
+	 *            type: integer
+	 *            format: int64
+	 *            minimum: 1
+	 *            readOnly: true
+	 *      responses:
+	 *        200:
+	 *          description: Successful operation
+	 *          content:
+	 *            application/json:
+	 *              schema:
+	 *                type: boolean
+	 */
+	private async delete(
+		options: APIHandlerOptions<{
+			params: { id: number };
+		}>,
+	): Promise<APIHandlerResponse> {
+		return {
+			payload: await this.commentService.delete(options.params.id),
+			status: HTTPCode.OK,
 		};
 	}