+++ apps/frontend/src/libs/modules/notification/libs/enums/notification-message.enum.ts
@@ -8,8 +8,11 @@ const NotificationMessage = {
 	GROUP_CREATED: "Group was created.",
 	GROUP_DELETED: "Group was deleted.",
 	GROUP_PERMISSIONS_CHANGED: "Group permissions were changed.",
+	PASSWORD_WAS_UPDATED: "Your password was updated.",
 	PROFILE_CHANGES_SAVED: "Changes are saved.",
 	SUBSCRIPTION_SUCCESS: "Congratulations! You are now a Premium Member 🎉",
+	USER_DELETED: "User was deleted.",
+	USER_DELETION_FAILED: "User wasn't deleted.",
 	USER_GROUPS_CHANGED: "User groups were changed.",
 } as const;