+++ apps/frontend/src/pages/management/libs/enums/management-dialogue-message.enum.ts
@@ -1,6 +1,7 @@
 const ManagementDialogueMessage = {
 	DELETE_COURSE: "Are you sure, that you want to delete course",
 	DELETE_GROUP: "Are you sure, that you want to delete group",
+	DELETE_USER: "Are you sure, that you want to delete user",
 } as const;
 
 export { ManagementDialogueMessage };